backend/requirements.txt
google-api-core==1.33.1
google-auth==1.35.0
google-cloud-bigquery==3.5.0
google-cloud-core==1.7.3
google-cloud-texttospeech==2.12.3
google-cloud-translate==2.0.1
pandas==1.5.0
fastapi
uvicorn